requirements.txt
Flask
Flask-SQLAlchemy
reportlab
arabic_reshaper
python-bidi
waitress
psycopg2-binary
pandas
openpyxl
Flask-Login
Werkzeug
email-validator
supabase==0.7.0
python-dotenv==0.19.2